refactor(diagnostics): move diagnostics packages to runtime
Re-home the diagnostics cluster from node.diagnostics to runtime.diagnostics without changing behavior.

Update dependent imports, package-local tests, and the SpotBugs exclusion for DefaultThreadDiagnostics.
