+++ build-logic/spotbugs-exclude.xml
@@ -1370,7 +1370,7 @@
   </Match>
   <Match>
     <Bug pattern="EI_EXPOSE_REP,EI_EXPOSE_REP2"/>
-    <Class name="network.crypta.node.diagnostics.threads.DefaultThreadDiagnostics"/>
+    <Class name="network.crypta.runtime.diagnostics.threads.DefaultThreadDiagnostics"/>
   </Match>
   <Match>
     <Bug pattern="EI_EXPOSE_REP,EI_EXPOSE_REP2"/>
