+++ src/main/java/network/crypta/runtime/admin/LegacyDiagnosticPort.java
@@ -30,14 +30,14 @@ import network.crypta.node.PeerManager;
 import network.crypta.node.PeerNodeStatus;
 import network.crypta.node.RequestTracker;
 import network.crypta.node.Version;
-import network.crypta.node.diagnostics.NodeDiagnostics;
-import network.crypta.node.diagnostics.ThreadDiagnostics;
-import network.crypta.node.diagnostics.threads.NodeThreadInfo;
-import network.crypta.node.diagnostics.threads.NodeThreadSnapshot;
 import network.crypta.node.stats.DataStoreInstanceType;
 import network.crypta.node.stats.DataStoreStats;
 import network.crypta.node.stats.StatsNotAvailableException;
 import network.crypta.node.stats.StoreAccessStats;
+import network.crypta.runtime.diagnostics.NodeDiagnostics;
+import network.crypta.runtime.diagnostics.ThreadDiagnostics;
+import network.crypta.runtime.diagnostics.threads.NodeThreadInfo;
+import network.crypta.runtime.diagnostics.threads.NodeThreadSnapshot;
 import network.crypta.runtime.spi.DiagnosticPort;
 import network.crypta.runtime.spi.DiagnosticReportSnapshot;
 import network.crypta.runtime.spi.DiagnosticSectionSnapshot;
