+++ src/test/java/network/crypta/node/subsystem/NodeServicesSubsystemTest.java
@@ -9,7 +9,6 @@ import network.crypta.config.SubConfig;
 import network.crypta.node.Node;
 import network.crypta.node.NodeClientCore;
 import network.crypta.node.NodeInitException;
-import network.crypta.node.diagnostics.DefaultNodeDiagnostics;
 import network.crypta.runtime.alerts.JVMVersionAlert;
 import network.crypta.runtime.alerts.MeaningfulNodeNameUserAlert;
 import network.crypta.runtime.alerts.NotEnoughNiceLevelsUserAlert;
@@ -18,6 +17,7 @@ import network.crypta.runtime.alerts.SimpleUserAlert;
 import network.crypta.runtime.alerts.TimeSkewDetectedUserAlert;
 import network.crypta.runtime.alerts.UserAlert;
 import network.crypta.runtime.alerts.UserAlertManager;
+import network.crypta.runtime.diagnostics.DefaultNodeDiagnostics;
 import network.crypta.runtime.updater.NodeUpdateManager;
 import network.crypta.support.JVMVersion;
 import network.crypta.support.PriorityAwareExecutor;
