+++ src/test/java/network/crypta/runtime/admin/LegacyDiagnosticPortTest.java
@@ -14,13 +14,13 @@ import network.crypta.node.NodeClientCore;
 import network.crypta.node.NodeStats;
 import network.crypta.node.PeerManager;
 import network.crypta.node.RequestTracker;
-import network.crypta.node.diagnostics.threads.NodeThreadInfo;
-import network.crypta.node.diagnostics.threads.NodeThreadSnapshot;
 import network.crypta.node.stats.DataStoreInstanceType;
 import network.crypta.node.stats.DataStoreKeyType;
 import network.crypta.node.stats.DataStoreStats;
 import network.crypta.node.stats.DataStoreType;
 import network.crypta.node.stats.StoreAccessStats;
+import network.crypta.runtime.diagnostics.threads.NodeThreadInfo;
+import network.crypta.runtime.diagnostics.threads.NodeThreadSnapshot;
 import network.crypta.runtime.spi.DiagnosticReportSnapshot;
 import network.crypta.runtime.spi.DiagnosticSectionSnapshot;
 import network.crypta.support.BandwidthStatsContainer;
